Remove default GitRepo Type from specification
Our OpenAPI validation library added support for default values. This
includes a feature/bug that fails the validation if defaults are set for
`merge-patch+json` requests.

We currently define a GitRepo type default of `auto` . This is defined
across all requests and responses. However this is arguably wrong for
patches. When patching a resource we generally don't want to set default
as this will lead to unexpected results (i.e. reverting to the default
if we do not explicitly set it to the current value in the patch)

For this reason (and to fix the validation error) we remove the GitRepo
type default from the specification.